requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.12.2
docker==4.4.4
flake8==3.8.4
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==1.5
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.58.0
whitenoise==5.2.0
